+++ translations/deepin-appstore_en_AU.ts
@@ -1,0 +1,120 @@
+<?xml version="1.0" ?><!DOCTYPE TS><TS language="en_AU" version="2.1">
+<context>
+    <name>QObject</name>
+    <message>
+        <location filename="../src/ui/widgets/search_completion_window.cpp" line="115"/>
+        <location filename="../src/ui/widgets/search_completion_window.cpp" line="164"/>
+        <source>Search &quot;%1&quot; in Deepin Store</source>
+        <translation>Search for &quot;%1&quot; in Deepin Store</translation>
+    </message>
+    <message>
+        <location filename="../src/ui/widgets/title_bar.cpp" line="201"/>
+        <source>Search</source>
+        <translation>Search</translation>
+    </message>
+    <message>
+        <location filename="../src/ui/widgets/title_bar_menu.cpp" line="68"/>
+        <source>Recommend App</source>
+        <translation>Recommend app</translation>
+    </message>
+    <message>
+        <location filename="../src/ui/widgets/title_bar_menu.cpp" line="74"/>
+        <source>Select region</source>
+        <translation type="unfinished"/>
+    </message>
+    <message>
+        <location filename="../src/ui/widgets/title_bar_menu.cpp" line="80"/>
+        <source>China</source>
+        <translation>China</translation>
+    </message>
+    <message>
+        <location filename="../src/ui/widgets/title_bar_menu.cpp" line="82"/>
+        <source>International</source>
+        <translation>International</translation>
+    </message>
+    <message>
+        <location filename="../src/ui/widgets/title_bar_menu.cpp" line="99"/>
+        <source>Clear cache</source>
+        <translation type="unfinished"/>
+    </message>
+    <message>
+        <location filename="../src/ui/widgets/title_bar_menu.cpp" line="103"/>
+        <source>Dark theme</source>
+        <translation type="unfinished"/>
+    </message>
+    <message>
+        <location filename="../src/ui/web_event_delegate.cpp" line="73"/>
+        <source>Undo</source>
+        <translation>Undo</translation>
+    </message>
+    <message>
+        <location filename="../src/ui/web_event_delegate.cpp" line="78"/>
+        <source>Redo</source>
+        <translation>Redo</translation>
+    </message>
+    <message>
+        <location filename="../src/ui/web_event_delegate.cpp" line="84"/>
+        <source>Cut</source>
+        <translation>Cut</translation>
+    </message>
+    <message>
+        <location filename="../src/ui/web_event_delegate.cpp" line="89"/>
+        <location filename="../src/ui/web_event_delegate.cpp" line="113"/>
+        <source>Copy</source>
+        <translation>Copy</translation>
+    </message>
+    <message>
+        <location filename="../src/ui/web_event_delegate.cpp" line="94"/>
+        <source>Paste</source>
+        <translation>Paste</translation>
+    </message>
+    <message>
+        <location filename="../src/ui/web_event_delegate.cpp" line="99"/>
+        <source>Delete</source>
+        <translation>Delete</translation>
+    </message>
+    <message>
+        <location filename="../src/ui/web_event_delegate.cpp" line="105"/>
+        <source>Select all</source>
+        <translation>Select all</translation>
+    </message>
+    <message>
+        <location filename="../src/app/deepin_appstore.cpp" line="107"/>
+        <source>Deepin Store</source>
+        <translation>Deepin Store</translation>
+    </message>
+    <message>
+        <location filename="../src/app/deepin_appstore.cpp" line="108"/>
+        <source>Deepin Store is an App Store with diverse and quality applications. It features popular recommendations, newly updated apps and hot topics, and supports one-click installation, updating and uninstalling.</source>
+        <translation type="unfinished"/>
+    </message>
+</context>
+<context>
+    <name>dstore::UserMenu</name>
+    <message>
+        <location filename="../src/ui/widgets/user_menu.cpp" line="13"/>
+        <source>Username</source>
+        <translation>Username</translation>
+    </message>
+    <message>
+        <location filename="../src/ui/widgets/user_menu.cpp" line="16"/>
+        <source>My comments</source>
+        <translation>My comments</translation>
+    </message>
+    <message>
+        <location filename="../src/ui/widgets/user_menu.cpp" line="20"/>
+        <source>My donations</source>
+        <translation>My donations</translation>
+    </message>
+    <message>
+        <location filename="../src/ui/widgets/user_menu.cpp" line="24"/>
+        <source>My apps</source>
+        <translation>My apps</translation>
+    </message>
+    <message>
+        <location filename="../src/ui/widgets/user_menu.cpp" line="30"/>
+        <source>Sign out</source>
+        <translation>Sign out</translation>
+    </message>
+</context>
+</TS>